move task lane config classes to core Druid
