+++ extensions-contrib/kubernetes-overlord-extensions/src/main/java/org/apache/druid/k8s/overlord/KubernetesOverlordModule.java
@@ -42,6 +42,8 @@ import org.apache.druid.indexing.common.tasklogs.FileTaskLogs;
 import org.apache.druid.indexing.overlord.RemoteTaskRunnerFactory;
 import org.apache.druid.indexing.overlord.TaskRunnerFactory;
 import org.apache.druid.indexing.overlord.WorkerTaskRunner;
+import org.apache.druid.indexing.overlord.config.TaskLane;
+import org.apache.druid.indexing.overlord.config.TaskLaneConfig;
 import org.apache.druid.indexing.overlord.config.TaskQueueConfig;
 import org.apache.druid.indexing.overlord.hrtr.HttpRemoteTaskRunnerFactory;
 import org.apache.druid.initialization.DruidModule;
@@ -49,9 +51,6 @@ import org.apache.druid.java.util.common.StringUtils;
 import org.apache.druid.java.util.common.lifecycle.Lifecycle;
 import org.apache.druid.java.util.common.logger.Logger;
 import org.apache.druid.k8s.overlord.common.DruidKubernetesClient;
-import org.apache.druid.k8s.overlord.common.TaskLane;
-import org.apache.druid.k8s.overlord.common.TaskLaneCapacityPolicy;
-import org.apache.druid.k8s.overlord.common.TaskLaneConfig;
 import org.apache.druid.k8s.overlord.common.TaskLaneRegistry;
 import org.apache.druid.k8s.overlord.runnerstrategy.RunnerStrategy;
 import org.apache.druid.tasklogs.NoopTaskLogs;
@@ -59,13 +58,11 @@ import org.apache.druid.tasklogs.TaskLogKiller;
 import org.apache.druid.tasklogs.TaskLogPusher;
 import org.apache.druid.tasklogs.TaskLogs;
 
-import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Properties;
 import java.util.Set;
-import java.util.stream.Collectors;
 
 
 @LoadScope(roles = NodeRole.OVERLORD_JSON_NAME)
@@ -174,12 +171,12 @@ public class KubernetesOverlordModule implements DruidModule
     List<TaskLaneConfig> taskLanes = runnerConfig.getTaskLanes();
     Map<String, TaskLane> labelToTaskLanes = new HashMap<>();
     for (TaskLaneConfig taskLaneConfig : taskLanes) {
-      TaskLaneCapacityPolicy policy = TaskLaneCapacityPolicy.valueOf(taskLaneConfig.getPolicy());
-      Set<String> labelSet = Arrays.stream(taskLaneConfig.getLabels().split(",\\s*"))
-                                   .map(String::trim)
-                                   .filter(label -> !label.isEmpty())
-                                   .collect(Collectors.toSet());
-      TaskLane taskLane = new TaskLane(labelSet, taskLaneConfig.getCapacityRatio(), policy);
+      Set<String> labelSet = taskLaneConfig.getLabelSet();
+      TaskLane taskLane = new TaskLane(
+          labelSet,
+          taskLaneConfig.getCapacityRatio(),
+          taskLaneConfig.getTaskLaneCapacityPolicy()
+      );
       labelSet.forEach(label -> labelToTaskLanes.put(label, taskLane));
     }
