+++ indexing-service/src/main/java/org/apache/druid/indexing/overlord/config/TaskLane.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.druid.k8s.overlord.common;
+package org.apache.druid.indexing.overlord.config;
 
 import java.util.Set;
